Update QEMU in Linux environment to use v10 (#1144)
## Description

This change picks up the v10.0.0 version of QEMU for Linux environment,
because we need a timer feature from this version to support hafnium.

Windows version is left out because Chocolatey is not updated to produce
this version yet.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This change is tested locally and booted to Windows.

## Integration Instructions

N/A