Cleanup ArmPlatformLib usage for QEMU SBSA (#1213)
## Description

The current state of platform is using NULL instance of ArmPlatformLib
for the SEC phase and the real QEMU SBSA instance for PEI phase. This
will cause the incorrect hobs to be created and consumed by later
components.

This change moves the heavy lifting hob creation routine at the
constructor of ArmPlatformLib to PlatformLib, which is only used at PEI.
So that all components are using the same instance of ArmPlatformLib.

The platform debugger is also disabled due to lack of notification when
exception occurs.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This change is tested on QEMU SBSA and verified that changing the number
of cores from ArmPlatformLib can impact the number of cores in MP
services.

## Integration Instructions

N/A