Move test service and inter-partition service to Rust based secure partition (#1202)
## Description

FF-A feature repo recently integrated the rust based secure partition.
This change is made to integrate the rust based secure partition as an
extra secure partition entry, into SBSA platform.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested on SBSA platform and passed the FF-A test application.

## Integration Instructions

N/A