Disable SVE/SME on QemuSbsaPkg
When FF-A support was added to QemuSbsaPkg, it re-enabled
SVE and SME for TF-A and QEMU, which had been disabled because
Windows boot was broken with those.

Windows boot is again broken because of SVE/SME and a change
in ArmPlatformPkg consumed from edk2 that aligned the feature
register reserved field behavior with the spec. A Windows change
will be forthcoming, but in the mean time, SVE/SME needs to
be disabled again.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>