Add key interrupt so that it will reset the console mode (#1242)
## Description

If QEMU is interrupted with `Ctrl + C`, the QEMU will leave the odd
console mode enabled.

There used to be some logic handling the console mode restoration, but
that did not work with key interrupt case.

This change is updated to fix that case.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Verified that QEMU interrupted by keyboard no longer has odd console
mode after the change.

## Integration Instructions

N/A

---------

Signed-off-by: kuqin12 <42554914+kuqin12@users.noreply.github.com>
Co-authored-by: Joey Vagedes <joey.vagedes@gmail.com>