Adding spin_lock usage to ensure no time related MP issues (#1258)
## Description

Current code has used a global variable to store the next instruction
after the system returned from its initial SMI.

This is not being guarded by any atomic locks, which could cause the
system to hang if the other APs come to change that first.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This change fixed the hanging issue on pipeline.

## Integration Instructions

N/A