Fixing watchdog timer driver SBSA (#1285)
## Description

The current usage of WatchdogTimer (WDT) from MdeModulePkg in SBSA is
incorrect.

This change moves to use the ARM specific version of WDT support.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested on QEMU SBSA and verified system will reset through the
interrupt handler as expected.

## Integration Instructions

N/A